pyproject.toml
[project]
name = "notely"
version = "0.1.0"
description = "Transform multimodal lectures into beautiful, structured Markdown notes"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Archer", email = "coder.archer@gmail.com" }]
keywords = ["lecture", "notes", "markdown", "asr", "ocr", "multimodal", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "ffmpeg-python>=0.2.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pdfplumber>=0.10.0",  # PDF processing (MIT license)
    "pyyaml>=6.0",  # YAML parsing for template frontmatter
]

[project.optional-dependencies]
funasr = ["funasr>=1.0.0", "modelscope>=1.10.0"]
paddleocr = ["paddleocr>=2.7.0", "paddlepaddle>=2.5.0"]
whisper = ["openai-whisper>=20231117"]
pdf-fast = ["pymupdf>=1.23.0"]  # High-performance PDF (AGPL license)
document = ["python-docx>=1.0.0", "python-pptx>=0.6.0"]  # Future: Word/PPT support
dev = ["ruff>=0.1.0", "pytest>=7.0.0", "pytest-cov>=4.0.0"]
all = [
    "funasr>=1.0.0",
    "modelscope>=1.10.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "openai-whisper>=20231117",
]

[project.scripts]
notely = "notely.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notely"]

[tool.hatch.build.targets.sdist]
include = ["/src", "/README.md", "/LICENSE"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # Line length handled by formatter
    "B008",   # Do not perform function call in argument defaults
    "B019",   # lru_cache on methods (acceptable for our use case)
    "B904",   # Exception chaining (intentionally suppressed for ImportError)
    "B905",   # zip() without strict (not needed for our use case)
    "SIM108", # Use ternary operator (sometimes less readable)
    "RUF001", # Ambiguous unicode characters (intentional for Chinese)
    "RUF002", # Ambiguous unicode characters in docstrings
    "RUF003", # Ambiguous unicode characters in comments
    "RUF005", # List concatenation (more readable as-is)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
strict = true
